01-intro-llms/issue-prioritizer/requirements.txt
# Core
fastapi>=0.128.0
uvicorn>=0.34.0
pydantic>=2.12.0
pydantic-settings>=2.12.0
python-dotenv>=1.0.0

# LLM
openai>=2.15.0
litellm>=1.80.0

# Evaluations
openevals>=0.1.3
pytest>=8.3.0

# UI
streamlit>=1.53.0
requests>=2.32.0
